Separate files for media extractor from initial preprocess file
